+++ src/server/services/aiAgent/__tests__/execAgent.deviceToolPipeline.test.ts
@@ -126,7 +126,7 @@ vi.mock('@/server/modules/Mecha', () => {
   };
 });
 
-vi.mock('@/server/services/toolExecution/deviceGateway', () => ({
+vi.mock('@/server/services/deviceGateway', () => ({
   deviceGateway: {
     get isConfigured() {
       // Will be overridden per-test via vi.spyOn or re-mock
@@ -209,7 +209,7 @@ describe('AiAgentService.execAgent - device tool pipeline ()', () => {
   describe('deviceContext forwarded to createServerAgentToolsEngine', () => {
     it('should pass deviceContext when gateway is configured', async () => {
       // Override deviceGateway.isConfigured
-      const { deviceGateway } = await import('@/server/services/toolExecution/deviceGateway');
+      const { deviceGateway } = await import('@/server/services/deviceGateway');
       vi.spyOn(deviceGateway, 'isConfigured', 'get').mockReturnValue(true);
       mockQueryDeviceList.mockResolvedValue([
         { deviceId: 'dev-1', deviceName: 'My PC', platform: 'win32' },
@@ -230,7 +230,7 @@ describe('AiAgentService.execAgent - device tool pipeline ()', () => {
     });
 
     it('should not pass deviceContext when gateway is not configured', async () => {
-      const { deviceGateway } = await import('@/server/services/toolExecution/deviceGateway');
+      const { deviceGateway } = await import('@/server/services/deviceGateway');
       vi.spyOn(deviceGateway, 'isConfigured', 'get').mockReturnValue(false);
 
       mockGetAgentConfig.mockResolvedValue(createBaseAgentConfig());
@@ -245,7 +245,7 @@ describe('AiAgentService.execAgent - device tool pipeline ()', () => {
 
   describe('RemoteDevice systemRole override', () => {
     it('should override RemoteDevice systemRole with dynamic prompt when enabled by ToolsEngine', async () => {
-      const { deviceGateway } = await import('@/server/services/toolExecution/deviceGateway');
+      const { deviceGateway } = await import('@/server/services/deviceGateway');
       vi.spyOn(deviceGateway, 'isConfigured', 'get').mockReturnValue(true);
       mockQueryDeviceList.mockResolvedValue([
         { deviceId: 'dev-1', deviceName: 'My PC', platform: 'win32' },
@@ -279,7 +279,7 @@ describe('AiAgentService.execAgent - device tool pipeline ()', () => {
     });
 
     it('should NOT have RemoteDevice in manifestMap when gateway is not configured', async () => {
-      const { deviceGateway } = await import('@/server/services/toolExecution/deviceGateway');
+      const { deviceGateway } = await import('@/server/services/deviceGateway');
       vi.spyOn(deviceGateway, 'isConfigured', 'get').mockReturnValue(false);
 
       // ToolsEngine returns empty manifestMap (RemoteDevice disabled by enableChecker)
@@ -301,7 +301,7 @@ describe('AiAgentService.execAgent - device tool pipeline ()', () => {
 
   describe('toolExecutorMap gating on gatewayConfigured (regression for #13769)', () => {
     it('should mark local-system as client when gateway is NOT configured (standalone Electron)', async () => {
-      const { deviceGateway } = await import('@/server/services/toolExecution/deviceGateway');
+      const { deviceGateway } = await import('@/server/services/deviceGateway');
       vi.spyOn(deviceGateway, 'isConfigured', 'get').mockReturnValue(false);
 
       mockGetEnabledPluginManifests.mockReturnValue(
@@ -316,7 +316,7 @@ describe('AiAgentService.execAgent - device tool pipeline ()', () => {
     });
 
     it('should NOT mark local-system as client when gateway IS configured (cloud)', async () => {
-      const { deviceGateway } = await import('@/server/services/toolExecution/deviceGateway');
+      const { deviceGateway } = await import('@/server/services/deviceGateway');
       vi.spyOn(deviceGateway, 'isConfigured', 'get').mockReturnValue(true);
       mockQueryDeviceList.mockResolvedValue([
         { deviceId: 'dev-1', deviceName: 'My PC', platform: 'win32' },
@@ -348,7 +348,7 @@ describe('AiAgentService.execAgent - device tool pipeline ()', () => {
       mockGetEnabledPluginManifests.mockReturnValue(new Map([['my-stdio-mcp', stdioManifest]]));
       mockGetAgentConfig.mockResolvedValue(createBaseAgentConfig({ plugins: ['my-stdio-mcp'] }));
 
-      const { deviceGateway } = await import('@/server/services/toolExecution/deviceGateway');
+      const { deviceGateway } = await import('@/server/services/deviceGateway');
 
       // Gateway NOT configured → should mark as client
       vi.spyOn(deviceGateway, 'isConfigured', 'get').mockReturnValue(false);
@@ -374,7 +374,7 @@ describe('AiAgentService.execAgent - device tool pipeline ()', () => {
       // Remote Device proxy to the device registered with the gateway, never
       // back to the caller. (The Phase 6.4 clientRuntime=desktop
       // short-circuit that bypassed this gate was removed.)
-      const { deviceGateway } = await import('@/server/services/toolExecution/deviceGateway');
+      const { deviceGateway } = await import('@/server/services/deviceGateway');
       vi.spyOn(deviceGateway, 'isConfigured', 'get').mockReturnValue(true);
       mockQueryDeviceList.mockResolvedValue([
         { deviceId: 'dev-1', deviceName: 'Remote VM', platform: 'linux' },
@@ -402,7 +402,7 @@ describe('AiAgentService.execAgent - device tool pipeline ()', () => {
         meta: { title: 'Stdio' },
       };
 
-      const { deviceGateway } = await import('@/server/services/toolExecution/deviceGateway');
+      const { deviceGateway } = await import('@/server/services/deviceGateway');
       vi.spyOn(deviceGateway, 'isConfigured', 'get').mockReturnValue(true);
       mockQueryDeviceList.mockResolvedValue([
         { deviceId: 'dev-1', deviceName: 'Remote VM', platform: 'linux' },
