+++ src/server/services/aiAgent/__tests__/execAgent.resumeApproval.test.ts
@@ -118,7 +118,7 @@ vi.mock('@/server/modules/Mecha', () => ({
   }),
 }));
 
-vi.mock('@/server/services/toolExecution/deviceGateway', () => ({
+vi.mock('@/server/services/deviceGateway', () => ({
   deviceGateway: { isConfigured: false, queryDeviceList: vi.fn().mockResolvedValue([]) },
 }));
