+++ src/server/services/aiAgent/__tests__/execAgent.threadId.test.ts
@@ -124,7 +124,7 @@ vi.mock('@/server/modules/Mecha', () => ({
 }));
 
 // Mock deviceGateway
-vi.mock('@/server/services/toolExecution/deviceGateway', () => ({
+vi.mock('@/server/services/deviceGateway', () => ({
   deviceGateway: {
     isConfigured: false,
     queryDeviceList: vi.fn().mockResolvedValue([]),
