+++ src/server/services/deviceGateway/index.ts
