+++ src/server/services/toolExecution/serverRuntimes/localSystem.ts
@@ -1,6 +1,7 @@
 import { LocalSystemIdentifier, LocalSystemManifest } from '@lobechat/builtin-tool-local-system';
 
-import { deviceGateway } from '../deviceGateway';
+import { deviceGateway } from '@/server/services/deviceGateway';
+
 import { type ServerRuntimeRegistration } from './types';
 
 export const localSystemRuntime: ServerRuntimeRegistration = {
