+++ src/server/services/toolExecution/serverRuntimes/remoteDevice.ts
@@ -3,7 +3,8 @@ import {
   RemoteDeviceIdentifier,
 } from '@lobechat/builtin-tool-remote-device';
 
-import { deviceGateway } from '../deviceGateway';
+import { deviceGateway } from '@/server/services/deviceGateway';
+
 import { type ServerRuntimeRegistration } from './types';
 
 export const remoteDeviceRuntime: ServerRuntimeRegistration = {
